convert main
int main -> public static void main as we move from C++ to C#.
cout statements are converted to Console.WriteLine.
taking in variables is converted from cin to setting them equal to Console.ReadLine() which takes the user input.
program is called with Program.bico(binom, coeff) instead of bico(binom, coeff) as that is how it is done in the new language.